dev-requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --extra=dev --output-file=dev-requirements.txt pyproject.toml
#
application-properties==0.8.2
    # via pymarkdownlnt
asgiref==3.8.1
    # via
    #   django
    #   django-stubs
attrs==23.2.0
    # via
    #   flake8-annotations
    #   flake8-bugbear
    #   jsonschema
    #   referencing
black==24.4.2
    # via grocery_store (pyproject.toml)
certifi==2024.7.4
    # via requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via black
columnar==1.4.1
    # via pymarkdownlnt
cryptography==42.0.8
    # via social-auth-core
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==5.0.7
    # via
    #   django-stubs
    #   django-stubs-ext
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-spectacular
    #   grocery_store (pyproject.toml)
    #   social-auth-app-django
django-stubs==5.0.2
    # via grocery_store (pyproject.toml)
django-stubs-ext==5.0.2
    # via
    #   django-stubs
    #   grocery_store (pyproject.toml)
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.15.2
    # via
    #   djangorestframework-simplejwt
    #   drf-spectacular
    #   grocery_store (pyproject.toml)
djangorestframework-simplejwt==5.3.1
    # via djoser
djoser==2.2.3
    # via grocery_store (pyproject.toml)
drf-spectacular==0.27.2
    # via grocery_store (pyproject.toml)
flake8==7.1.0
    # via
    #   flake8-annotations
    #   flake8-bugbear
    #   flake8-commas
    #   flake8-pyproject
    #   grocery_store (pyproject.toml)
    #   pep8-naming
flake8-annotations==3.1.1
    # via grocery_store (pyproject.toml)
flake8-bugbear==24.4.26
    # via grocery_store (pyproject.toml)
flake8-commas==4.0.0
    # via grocery_store (pyproject.toml)
flake8-pyproject==1.2.3
    # via grocery_store (pyproject.toml)
idna==3.7
    # via requests
inflection==0.5.1
    # via drf-spectacular
isort==5.13.2
    # via grocery_store (pyproject.toml)
jsonschema==4.23.0
    # via drf-spectacular
jsonschema-specifications==2023.12.1
    # via jsonschema
mccabe==0.7.0
    # via flake8
mypy==1.10.1
    # via grocery_store (pyproject.toml)
mypy-extensions==1.0.0
    # via
    #   black
    #   grocery_store (pyproject.toml)
    #   mypy
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.1
    # via black
pathspec==0.12.1
    # via black
pep8-naming==0.14.1
    # via grocery_store (pyproject.toml)
pillow==10.4.0
    # via grocery_store (pyproject.toml)
platformdirs==4.2.2
    # via black
pycodestyle==2.12.0
    # via flake8
pycparser==2.22
    # via cffi
pyflakes==3.2.0
    # via flake8
pyjwt==2.8.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pymarkdownlnt==0.9.21
    # via grocery_store (pyproject.toml)
python3-openid==3.2.0
    # via social-auth-core
pyyaml==6.0.1
    # via
    #   application-properties
    #   drf-spectacular
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
rpds-py==0.19.0
    # via
    #   jsonschema
    #   referencing
social-auth-app-django==5.4.1
    # via djoser
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.0
    # via django
tomli==2.0.1
    # via application-properties
toolz==0.12.1
    # via columnar
types-pyyaml==6.0.12.20240311
    # via django-stubs
typing-extensions==4.12.2
    # via
    #   application-properties
    #   django-stubs
    #   django-stubs-ext
    #   mypy
    #   pymarkdownlnt
uritemplate==4.1.1
    # via drf-spectacular
urllib3==2.2.2
    # via requests
wcwidth==0.2.13
    # via columnar
